Allow to set filename only ID's
Currently either one completely parses the URI again and thus knows what extension to set in the ID field, or uses the default. There is no way to only set the filename and automatically set the extension. This added line will make sure, that the extension is appended, if the user only provided e.g. "myfancyname" as the ID.